backend/requirements.txt
# LexiLens Backend Dependencies
# Python 3.11+

# --- Web Framework ---
fastapi==0.115.8
uvicorn[standard]==0.34.0
websockets==14.2

# --- Google AI ---
google-genai==1.5.0
google-adk==1.2.1

# --- Google Cloud ---
google-cloud-secret-manager==2.22.0
google-cloud-logging==3.11.4
google-cloud-firestore==2.20.1

# --- Config & Validation ---
pydantic==2.10.6
pydantic-settings==2.8.1

# --- Utilities ---
python-dotenv==1.0.1
